refactor(dashboard): isolate operator health parsing
