+++ dashboard/src/guard-api.ts
@@ -8,7 +8,9 @@ import {
   CODEX_RESUME_STATUSES
 } from "./guard-types";
 import { computeTrendBuckets } from "./evidence/evidence-metrics";
+import { normalizeOperatorHealth } from "./operator-health";
 import { normalizeProtectionHealth, protectionHeadlineFor } from "./protection-health";
+export { normalizeOperatorHealth } from "./operator-health";
 import {
   AUTHORITATIVE_DECISION_INCONSISTENT,
   guardActionDisposition,
@@ -1644,38 +1646,6 @@ function normalizeCloudCommandCapability(raw: unknown): GuardRuntimeSnapshot["cl
       typeof raw["revoke_command"] === "string"
         ? raw["revoke_command"]
         : "hol-guard commands revoke --confirm revoke",
-  };
-}
-
-function nonNegativeNumber(value: unknown): number {
-  return typeof value === "number" && Number.isFinite(value) && value >= 0 ? value : 0;
-}
-
-export function normalizeOperatorHealth(raw: unknown): GuardRuntimeSnapshot["operator_health"] {
-  if (!isRecord(raw)) {
-    return undefined;
-  }
-  const state = raw["state"];
-  const cause = raw["cause"];
-  const automaticRecovery = raw["automatic_recovery"];
-  if (
-    !["healthy", "backlogged", "saturated", "store-contended"].includes(String(state))
-    || typeof cause !== "string"
-    || typeof automaticRecovery !== "string"
-  ) {
-    return undefined;
-  }
-  return {
-    state: state as "healthy" | "backlogged" | "saturated" | "store-contended",
-    cause,
-    automatic_recovery: automaticRecovery,
-    repairable: raw["repairable"] === true,
-    queue_depth: nonNegativeNumber(raw["queue_depth"]),
-    queue_limit: nonNegativeNumber(raw["queue_limit"]),
-    oldest_wait_ms: nonNegativeNumber(raw["oldest_wait_ms"]),
-    workers_busy: nonNegativeNumber(raw["workers_busy"]),
-    workers_ready: nonNegativeNumber(raw["workers_ready"]),
-    workers_configured: nonNegativeNumber(raw["workers_configured"]),
   };
 }
