+++ src/codex_plugin_scanner/guard/daemon/static/assets/guard-dashboard.js
@@ -15288,6 +15288,35 @@ function computePeriodComparison(receipts, days, now2) {
     totalDelta: currentTotal - previousTotal
   };
 }
+function nonNegativeNumber(value) {
+  return typeof value === "number" && Number.isFinite(value) && value >= 0 ? value : 0;
+}
+function isRecord$4(value) {
+  return typeof value === "object" && value !== null && !Array.isArray(value);
+}
+function normalizeOperatorHealth(raw) {
+  if (!isRecord$4(raw)) {
+    return void 0;
+  }
+  const state = raw["state"];
+  const cause = raw["cause"];
+  const automaticRecovery = raw["automatic_recovery"];
+  if (!["healthy", "backlogged", "saturated", "store-contended"].includes(String(state)) || typeof cause !== "string" || typeof automaticRecovery !== "string") {
+    return void 0;
+  }
+  return {
+    state,
+    cause,
+    automatic_recovery: automaticRecovery,
+    repairable: raw["repairable"] === true,
+    queue_depth: nonNegativeNumber(raw["queue_depth"]),
+    queue_limit: nonNegativeNumber(raw["queue_limit"]),
+    oldest_wait_ms: nonNegativeNumber(raw["oldest_wait_ms"]),
+    workers_busy: nonNegativeNumber(raw["workers_busy"]),
+    workers_ready: nonNegativeNumber(raw["workers_ready"]),
+    workers_configured: nonNegativeNumber(raw["workers_configured"])
+  };
+}
 const PROTECTION_CHECK_IDS = [
   "harness_hooks",
   "daemon",
@@ -16703,32 +16732,6 @@ function normalizeCloudCommandCapability(raw) {
     pending_commands: pending,
     enable_command: typeof raw["enable_command"] === "string" ? raw["enable_command"] : "hol-guard commands enable --operations read-only",
     revoke_command: typeof raw["revoke_command"] === "string" ? raw["revoke_command"] : "hol-guard commands revoke --confirm revoke"
-  };
-}
-function nonNegativeNumber(value) {
-  return typeof value === "number" && Number.isFinite(value) && value >= 0 ? value : 0;
-}
-function normalizeOperatorHealth(raw) {
-  if (!isRecord$1(raw)) {
-    return void 0;
-  }
-  const state = raw["state"];
-  const cause = raw["cause"];
-  const automaticRecovery = raw["automatic_recovery"];
-  if (!["healthy", "backlogged", "saturated", "store-contended"].includes(String(state)) || typeof cause !== "string" || typeof automaticRecovery !== "string") {
-    return void 0;
-  }
-  return {
-    state,
-    cause,
-    automatic_recovery: automaticRecovery,
-    repairable: raw["repairable"] === true,
-    queue_depth: nonNegativeNumber(raw["queue_depth"]),
-    queue_limit: nonNegativeNumber(raw["queue_limit"]),
-    oldest_wait_ms: nonNegativeNumber(raw["oldest_wait_ms"]),
-    workers_busy: nonNegativeNumber(raw["workers_busy"]),
-    workers_ready: nonNegativeNumber(raw["workers_ready"]),
-    workers_configured: nonNegativeNumber(raw["workers_configured"])
   };
 }
 function normalizeRuntimeSnapshot(snapshot) {
